requirements.txt
flask
requests
google-generativeai
PyGithub
gunicorn